+++ Server/server.hpp
@@ -11,7 +11,7 @@ class __server {
         std::string                 port;
         std::string                 server_name;
         std::map<std::string, std::string>  error_page;
-        size_t                      client_max_body_size;
+        unsigned long long          client_max_body_size;
         std::map <std::string, __location>    location;
 
     public:
@@ -22,7 +22,7 @@ class __server {
         std::string                         get_port();
         std::string                         get_server_name();
         std::map<std::string, std::string>  get_error_pages();
-        size_t                              get_client_max_body_size();
+        unsigned long long                  get_client_max_body_size();
         std::map <std::string, __location>  get_locations();
 
         int    set_host(std::string host);
